chore: Integrate FastAPI into main.py and update .gitignore
